backend/requirements.txt.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pinecone-client==2.2.4
openai==0.28.1
requests==2.31.0
pydantic==2.4.2
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==0.8.11
docx2txt==0.8
Pillow==10.1.0
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.25.2
typing-extensions==4.8.0